change Crab Site latitude and longitude

+++ ltserChanghua/data/models.py
@@ -37,8 +37,8 @@ class LatestEvent(models.Model):
 
 class CrabSite(models.Model):
     title = models.CharField(max_length=200)
-    latitude = models.DecimalField(max_digits=10, decimal_places=5, null=True, blank=True)
-    longitude = models.DecimalField(max_digits=10, decimal_places=5, null=True, blank=True)
+    latitude = models.FloatField(null=True, blank=True)
+    longitude =  models.FloatField(null=True, blank=True)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
     def __str__(self):
